Add validation rules for coordinates

+++ src/test/java/nl/tudelft/thefirstorder/web/rest/CameraResourceIntTest.java
@@ -45,11 +45,11 @@ public class CameraResourceIntTest {
     private static final String DEFAULT_NAME = "AAAAA";
     private static final String UPDATED_NAME = "BBBBB";
 
-    private static final Integer DEFAULT_X = 1;
-    private static final Integer UPDATED_X = 2;
+    private static final Double DEFAULT_X = 1D;
+    private static final Double UPDATED_X = 2D;
 
-    private static final Integer DEFAULT_Y = 1;
-    private static final Integer UPDATED_Y = 2;
+    private static final Double DEFAULT_Y = 1D;
+    private static final Double UPDATED_Y = 2D;
     private static final String DEFAULT_CAMERA_TYPE = "AAAAA";
     private static final String UPDATED_CAMERA_TYPE = "BBBBB";
     private static final String DEFAULT_LENS_TYPE = "AAAAA";
